Added temporarily backup for generate command
