+++ drizzle-kit/src/cli/commands/libSqlPushUtils.ts
@@ -11,6 +11,7 @@ import {
 	SQLiteDropTableConvertor,
 	SqliteRenameTableConvertor,
 } from '../../sqlgenerator';
+import { collectCascadeDependents } from '../../utils/cascade';
 
 export const getOldTableName = (
 	tableName: string,
@@ -123,31 +124,6 @@ export const _moveDataStatements = (
 	}
 
 	return statements;
-};
-
-const collectCascadeDependents = (
-	rootTable: string,
-	json: SQLiteSchemaSquashed,
-): string[] => {
-	const result = new Set<string>();
-	const queue = [rootTable];
-
-	while (queue.length) {
-		const current = queue.pop()!;
-		for (const table of Object.values(json.tables)) {
-			for (const fk of Object.values(table.foreignKeys)) {
-				const data = SQLiteSquasher.unsquashPushFK(fk);
-				if (data.tableTo === current && data.onDelete === 'cascade') {
-					if (!result.has(table.name)) {
-						result.add(table.name);
-						queue.push(table.name);
-					}
-				}
-			}
-		}
-	}
-
-	return Array.from(result);
 };
 
 export const libSqlLogSuggestionsAndReturn = async (
